PR 생성 시 서버 API 호출로 브랜치커밋 가이드 자동 댓글 : refactor : 기본 api 골격 및 config를 WEB모듈로 이동, Entity, repository 도 각 모듈로 리펙토링 github관련 모듈 리펙토링 https://github.com/Cassiiopeia/suh-project-utility/issues/48
