+++ Suh-Domain-Github/src/main/java/me/suhsaechan/github/entity/GithubIssueHelper.java
@@ -1,4 +1,4 @@
-package me.suhsaechan.common.entity;
+package me.suhsaechan.github.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,14 +7,15 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
-
-import java.util.UUID;
+import me.suhsaechan.common.entity.BasePostgresEntity;
+import me.suhsaechan.github.repository.GithubRepository;
 
 @Entity
 @Getter
