+++ Suh-Web/src/main/java/me/suhsaechan/web/config/ServerInfo.java
@@ -1,4 +1,4 @@
-package me.suhsaechan.common.config;
+package me.suhsaechan.web.config;
 
 import lombok.Getter;
 import lombok.Setter;
