+++ Suh-Web/src/main/java/me/suhsaechan/web/controller/IssueHelperController.java
@@ -2,6 +2,8 @@ package me.suhsaechan.web.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import me.suhsaechan.github.dto.IssueHelperRequest;
+import me.suhsaechan.github.dto.IssueHelperResponse;
 import me.suhsaechan.github.service.IssueHelperService;
 import me.suhsaechan.suhlogger.annotation.LogMonitor;
 import org.springframework.http.MediaType;
@@ -21,17 +23,18 @@ public class IssueHelperController {
 
   @LogMonitor
   @PostMapping(value = "/create/commmit-branch", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-  public ResponseEntity<me.suhsaechan.github.dto.response.IssueHelperResponse> createIssueCommmitBranch(
-      @ModelAttribute me.suhsaechan.github.dto.request.IssueHelperRequest request) {
-    try {
-      me.suhsaechan.github.dto.response.IssueHelperResponse response = issueHelperService.createIssueCommmitBranch(request);
+  public ResponseEntity<IssueHelperResponse> createIssueCommmitBranch(
+      @ModelAttribute IssueHelperRequest request) {
+      IssueHelperResponse response = issueHelperService.createIssueCommmitBranch(request);
       return ResponseEntity.ok(response);
-    } catch (IllegalArgumentException e) {
-      log.error("입력 오류: {}", e.getMessage());
-      return ResponseEntity.badRequest().build();
-    } catch (Exception e) {
-      log.error("서버 오류: {}", e.getMessage());
-      return ResponseEntity.internalServerError().build();
-    }
+  }
+
+  // GITHUB_WORKFLOW 에서 호출되는 API
+  @LogMonitor
+  @PostMapping(value = "/create/commmit-branch/github-workflow", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+  public ResponseEntity<IssueHelperResponse> createIssueCommmitBranchByGithubWorkflow(
+      @ModelAttribute IssueHelperRequest request) {
+      IssueHelperResponse response = issueHelperService.createIssueCommmitBranch(request);
+      return ResponseEntity.ok(response);
   }
 }
