+++ Suh-Web/src/main/java/me/suhsaechan/web/controller/PageController.java
@@ -3,7 +3,7 @@ package me.suhsaechan.web.controller;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.suhsaechan.common.config.UserAuthority;
+import me.suhsaechan.web.config.UserAuthority;
 import me.suhsaechan.notice.service.NoticeService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
